[ADD]: Image class

+++ source/utils/Vec2.cpp
@@ -4,7 +4,7 @@ Vec2::Vec2(void) {
 	x = y = 0;
 }
 
-Vec2::Vec2(int x, int y) {
+Vec2::Vec2(float x, float y) {
 	this->x = x;
 	this->y = y;
 }
